+++ scripts/Groovy/lib/entities/Query.groovy
@@ -8,6 +8,9 @@ class Query {
     String from
     List<Reduction> reductions
     List<Transformation> transformations
-    List transformers = []
     CompoundFilter where
+
+    // Internal fields. Will be ignored even if specified in the query
+    List transformers = []
+    Map columns = [:]
 }